requirements.txt
aiofiles==0.8.0
aiogram @ https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
magic-filter==1.0.6
multidict==6.0.2
psycopg2-binary==2.9.3
pydantic==1.9.0
Pygments==2.11.2
python-dotenv==0.20.0
pytz==2022.1
typing-extensions==4.1.1
yarl==1.7.2
